+++ reana-ui/src/components/App.js
@@ -29,6 +29,7 @@ import Profile from "~/pages/profile/Profile";
 import PrivacyNotice from "~/pages/privacyNotice/PrivacyNotice";
 import Status from "~/pages/status/Status";
 import LaunchOnReana from "~/pages/launchOnReana/LaunchOnReana";
+import NotFound from "~/pages/error/NotFound";
 import Error from "./Error";
 
 import "./App.module.scss";
@@ -86,6 +87,8 @@ export default function App() {
           <ProtectedRoute path="/profile" component={Profile} />
           <ProtectedRoute path="/status" component={Status} />
           <ProtectedRoute exact path="/launch" component={LaunchOnReana} />
+          <ProtectedRoute path="/404" component={NotFound} />
+          <Redirect to="/404" />
         </Switch>
       )}
     </BrowserRouter>